feat: split sdk connect
