+++ app/core/SDKConnect/utils/wait.util.ts
@@ -1,8 +1,9 @@
 import { KeyringController } from '@metamask/keyring-controller';
 import { AndroidClient } from '../AndroidSDK/android-sdk-types';
 import RPCQueueManager from '../RPCQueueManager';
-import { Connection, SDKConnect } from '../SDKConnect';
+import { SDKConnect } from '../SDKConnect';
 import DevLogger from './DevLogger';
+import { Connection } from '../Connection';
 
 export const MAX_QUEUE_LOOP = Infinity;
 export const wait = (ms: number) =>
